requirements.txt
asgiref
certifi
chardet
charset-normalizer
cloudinary
crispy-bootstrap5
dj-database-url
dj-rest-auth
Django
django-allauth
django-cloudinary-storage
django-cors-headers
django-crispy-forms
django-environ
django-filter
djangorestframework
gunicorn
idna
Markdown
MouseInfo
packaging
pillow
psycopg2-binary
PyAutoGUI
PyGetWindow
PyMsgBox
pyperclip
PyRect
PyScreeze
pytweening
pytz
reportlab
requests
setuptools
six
sqlparse
typing_extensions
tzdata
urllib3
wheel
whitenoise